windows-clang: scrape DEFINE_KNOWN_FOLDER, restore the UIA _Impl scaffolding
knownfolders.h defines every FOLDERID_* through DEFINE_KNOWN_FOLDER, which
expands to the same EXTERN_C const GUID declaration as DEFINE_GUID and takes the
same twelve arguments — but the macro name was not recognised, so the header
yielded no metadata at all and every known-folder GUID was dropped. Without
INITGUID these expand to a valueless extern declaration, so the value lives only
in the macro arguments; parsing them is what makes the constants projectable.

Recognising the macro recovers 142 known folders (FOLDERID_ProgramData's GUID
matches the SDK header exactly) and gives the header its own partition.
Regenerating the package crates also surfaces the previously-added
dispatcherqueue and uiautomationcoreapi headers, which had reached the metadata
but not yet the published projections.

Separately, the --implement list still named the retired sub-namespaces
(Windows.Win32.UI.Accessibility.*), but the canonical Win32 metadata is a flat
Windows.Win32 namespace — the per-header partition is synthesised only for the
published package crates. A stale sub-namespace matches nothing silently, so no
_Impl scaffolding was emitted at all and the DirectComposition backend could not
implement any UIA provider. Naming the interfaces directly under Windows.Win32,
as system.txt already does, takes the emitted scaffolding from 8 to 182.

tool_roundtrip passes.

+++ crates/libs/reactor/src/backend/dcomp/host.rs
@@ -277,7 +277,7 @@ impl DCompHost {
 
         render_host.kick();
         unsafe {
-            let _ = ShowWindow(hwnd, SW_SHOW);
+            let _ = ShowWindow(hwnd, SW_SHOW as i32);
         }
         Ok(Self { hwnd })
     }
@@ -563,7 +563,7 @@ fn track_leave(hwnd: HWND) {
 fn frame_y_px(hwnd: HWND) -> i32 {
     let dpi = effective_dpi(hwnd);
     unsafe {
-        GetSystemMetricsForDpi(SM_CYFRAME, dpi) + GetSystemMetricsForDpi(SM_CXPADDEDBORDER, dpi)
+        GetSystemMetricsForDpi(SM_CYFRAME as i32, dpi) + GetSystemMetricsForDpi(SM_CXPADDEDBORDER as i32, dpi)
     }
 }
